med-verify-scan-main/backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-jwt-extended>=4.5.0
psycopg2-binary>=2.9.0
opencv-python>=4.8.0
pyzbar>=0.1.9
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
cryptography>=41.0.0
web3>=5.31.3
pytesseract>=0.3.10
tensorflow>=2.13.0
numpy>=1.24.0
celery>=5.3.0
redis>=5.0.0
apscheduler>=3.10.0
python-dateutil>=2.8.0
openai>=1.0.0
sendgrid>=6.10.0
pyfcm>=1.5.0
twilio>=8.10.0
werkzeug>=2.3.0
python-multipart>=0.0.6
pytz>=2023.3
scikit-learn>=1.3.0
qrcode[pil]>=8.0.0
reportlab>=4.0.0

